Add files via upload
Objective:
EECS 355 -- Assignment 2 -- Combinational Divider
In this lab, you will design a divider that operates on unsigned numbers. In this portion of the lab, you will construct the core portion of the divider using a combinational circuit (made out of a network of multiple combinational comparators). This circuit will then be governed with a start button control signal.
Almost everything used in this lab will be used again in Assignment 3. Keep everything neat and organized so that it can be easily reused. (REUSE: Golden rule of design!)
Background Information:
For the divider, you will use the long-division algorithm.
See http://en.wikipedia.org/wiki/Long_division if you need to refresh yourself on the long-division method. The same procedure can also be applied to binary numbers. Before continuing with the lab, try to perform a binary long division of 110001011 and 0101. Please include this division in your report.
The process of long division, is really about subtraction. Each bit of the quotient comes from the result of the subtraction of the previous bit. In a combinational divider, each of these subtractors is instantiated as a separate unit and data is passed from one to the next, on down the chain. The advantage of this method is that the calculation in a given stage can occur as soon as data is ready at the input. This is because each subtractor (also referred to in this lab as a comparator) can proceed with its calculation as soon as the previous one has finished.