Add files via upload
Objective:
EECS 355 – Assignment 1 Calculator
In this lab you will design a calculator that is capable of performing 1. addition
2. subtraction
3. multiplication
of two signed numbers. This calculator will be implemented on the Altera DE2-115 development kits.
Lab Description: I. Create new project:
1) Open a new blank project in Quartus II. The top level entity should be named something like “display_calc” because this project will involve both a calculator entity and an LED display module as sub-modules.
2) Assemble the LED decoder design from Lab 1 into a package called “decoder”.
How to create a package:
A package is essentially VHDL code saved as a .vhd file containing the definitions of user defined data types, constants, functions, and component declarations of design modules, which you have designed earlier and would like to re-use in future designs. For example we will reuse the LED decoder that was designed in Lab1 as a component in Lab2 as part of the calculator.