Add files via upload
EECS 355 – Assignment 4 Behavioral FSM Divider
In this lab, you will design a behavioral divider that operates on signed numbers. The divider will be designed using a Finite State Machine, which operates under the control of a start button.
Background Information:
The theory behind the behavioral divider is based on shift & subtract. Normally, this iterative process would take linear O(n) time in the number of bits. However, we can reduce the number of iterations significantly to O(log n) if we can predetermine the number of bits to shift in each iteration. We accomplish this by identifying the difference in position between the most significant bits of the dividend and divisor.
In each iteration, we calculate the difference in most significant bits between A and B, defined as P. In the sudo-function below, the function get_msb_pos() is called to get the top most bit. We then shift B by P bits to calculate the new A in each iteration. Note that if B << P is greater than A, then we must reduce P by one so that the result of the subtraction is a positive number. The quotient, Q, is the summation of 2P (or, 1 << P) values from each iteration.
In this lab, you will instantiate the divider using a finite state machine. In doing so, you should be implementing the while loop as a state machine with a branch back. DO NOT USE A WHILE LOOP IN PLACE OF YOUR STATE MACHINE.